improve some more stuff api-wise (#4411)
* just some docs for chaos.rs

* improve chaos apis a bit
